+++ backend/core_db_entities/src/entity/product_images.rs
@@ -7,9 +7,9 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(table_name = "ProductImages")]
 pub struct Model {
     #[sea_orm(column_name = "ImageID", primary_key)]
-    pub image_id: i32,
+    pub image_id: i64,
     #[sea_orm(column_name = "ProductID")]
-    pub product_id: Option<i32>,
+    pub product_id: Option<i64>,
     #[sea_orm(column_name = "ImageURL")]
     pub image_url: Option<String>,
     #[sea_orm(column_name = "AltText")]
